feat: remove `@clr/ui` shim (moving it to the `@clr/ui` package)
This commit removes the shim from `@cds/core`.

We are moving the shim to the `@clr/ui` package to allow both to be
evolved together. Otherwise, changes to the custom property API exposed
by `@clr/ui` will require changes in a separate package (and separate repo).
This would cause version synchronization difficulties.

This change is technically a breaking one, but we are making an
exception to semver for this change.

How to update:

The file path in applications should be updated as follows:
'@cds/core/styles/shim.clr-ui.min.css' -> '@clr/ui/shim.cds-core.min.css'

That's it, just a simple find and replace.