fix(react): fix jsdom react testing with vitest
Replace isJestTest with more general isJsdomTest.

The code had a workaround to ensure compatibility with Jest + Jsdom. The
value of `process.env.JEST_WORKER_ID` was used to determine whether the
execution environment was Jest. To ensure compatibility with other test
frameworks, and in particular Vitest, we use a more general test to
determine whether to apply the workaround.

Fixes #260

Signed-off-by: Antonin Bas <abas@vmware.com>