feat(file): allow custom button label (#295)
## What is the current behavior?

The `buttonLabel` property on `cds-file` can be set even though it is
private. When the file selection is cleared, the button label is reset
to "Browse" instead of the previous `buttonLabel` value.

Issue Number: CDE-1055

## What is the new behavior?

A custom button label is fully supported.

## Does this PR introduce a breaking change?

No.

## Other Information

- The `buttonLabel` property is now `public`.
- The is a new property named `buttonLabelForSelection`. This property
is `private` and should not be set externally.